Changed parallelStream to stream
The parallelStream isn't working because all the new threads created
for the job aren't in the same transactional context as the original one.
Also the current implementation of parallelStream is pretty bad. You can
read more about it (https://dzone.com/articles/think-twice-using-java-8).
I would like to cite "The problem is that all parallel streams use
common fork-join thread pool, and if you submit a long-running task,
you effectively block all threads in the pool.". So yes, we shouldn't
use parallelStream for now.